limit resources to a max of 10, move out tests to a diff file, never get all entitlements
